Feature/#46 admin review like system

+++ cookie/src/main/java/com/cookie/domain/review/dto/request/CreateReviewRequest.java
@@ -14,7 +14,7 @@ import lombok.NoArgsConstructor;
 public class CreateReviewRequest {
     private long movieId;
     private String content;
-    private double movieScore;
+    private Integer movieScore;
     private boolean isSpoiler;
 
     public Review toEntity(User user, Movie movie) {
